finish up to deploy , thank you MERN

+++ Ecommerce/controllers/categoryCtrl.js
@@ -1,5 +1,5 @@
 const Category = require('../models/categoryModel')
-
+const Products = require('../models/productModel')
 
 const categoryCtrl = {
 
@@ -36,6 +36,12 @@ const categoryCtrl = {
     },
     deleteCategory: async (req,res)=>{
         try{
+            const products = await Products.findOne({category:req.params.id})
+            if(products){
+                return res.status(400).json({
+                    msg:"Please delete all products with a relationship."
+                })
+            }
             await Category.findByIdAndDelete(req.params.id)
             res.json({msg:"Deleted a category"})
         }catch (err){
